+++ Student.Core.API/Migrations/myDbContextModelSnapshot.cs
@@ -111,9 +111,9 @@ namespace Student.Core.API.Migrations
 
             modelBuilder.Entity("Student.Model.StudentInfo", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int")
+                        .HasColumnType("bigint")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<string>("Address")
@@ -160,7 +160,7 @@ namespace Student.Core.API.Migrations
                     b.HasData(
                         new
                         {
-                            Id = 1,
+                            Id = 1L,
                             Address = "朝阳区朝阳公园西路9号院九号",
                             DepartId = 4,
                             Email = "xiaoan@stu.com",
@@ -175,7 +175,7 @@ namespace Student.Core.API.Migrations
                         },
                         new
                         {
-                            Id = 2,
+                            Id = 2L,
                             Address = "北京市朝阳区东三环中路甲10号",
                             DepartId = 3,
                             Email = "laoli@stu.com",
